bugfix(remove friend): Fixed bug where removing friend would not work if they added you
Changes made by Oliver Wuttke, Bachelor of Computer Science (Artificial Intelligence), WUTT0019.

+++ src/user/friends/remove_friend.php
@@ -12,11 +12,11 @@ require_once '../../inc/functions.php';
 $uid = get_uid($_SESSION['username']);
 $fid = $_POST['friend_id'];
 
-$sql = 'DELETE FROM Friendships WHERE user_id = ? AND friend_id = ?;';
-$statement = $conn->prepare($sql);
-$statement->bind_param('ii', $uid, $fid);
-$statement->execute();
-$statement->close();
+$sql = 'DELETE FROM Friendships WHERE (user_id = ? AND friend_id = ?) OR (user_id = ? AND friend_id = ?)';
+$stmt = $conn->prepare($sql);
+$stmt->bind_param('iiii', $uid, $fid, $fid, $uid);
+$stmt->execute();
+$stmt->close();
 
 header('Content-Type: application/json');
 echo json_encode(['success' => true]);
